Refactor level result screen and node preview
Remove deprecated level ui scripts
