+++ Assets/AddressableAssetsData/AssetGroups/UI.asset
@@ -37,6 +37,11 @@ MonoBehaviour:
     m_ReadOnly: 0
     m_SerializedLabels: []
     FlaggedDuringContentUpdateRestriction: 0
+  - m_GUID: a5fec944b72284587b727de4def7f144
+    m_Address: LevelResultScreen
+    m_ReadOnly: 0
+    m_SerializedLabels: []
+    FlaggedDuringContentUpdateRestriction: 0
   - m_GUID: fcd1972f74b934ba9ba1a31b11299c34
     m_Address: LevelUpResultScreen
     m_ReadOnly: 0
