+++ Assets/Prefabs/UI/Screens/LevelResultScreen.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3e170cfe8607846f8b50136c7113588e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
