+++ Assets/Prefabs/UI/Screens/LevelResultScreen/LevelResultScreen.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a5fec944b72284587b727de4def7f144
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
