+++ Assets/Prefabs/UI/Screens/LevelResultScreen/LevelResultScreenController.overrideController
@@ -1,0 +1,15 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!221 &22100000
+AnimatorOverrideController:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: LevelResultScreenController
+  m_Controller: {fileID: 9100000, guid: 9b97e147ef709f143879d238521ab680, type: 2}
+  m_Clips:
+  - m_OriginalClip: {fileID: 7400000, guid: 955558bf69698b2449ec435993f6b105, type: 2}
+    m_OverrideClip: {fileID: 7400000, guid: 0bfb8416bd6fb401fb9eaadaf2a3826f, type: 2}
+  - m_OriginalClip: {fileID: 7400000, guid: 3cb89df71c561834cb2c17e182adc89d, type: 2}
+    m_OverrideClip: {fileID: 7400000, guid: b88b7759fe9c040679d7f41c711e04d8, type: 2}
