+++ Assets/Prefabs/UI/Screens/LevelResultScreen/LevelResultScreenController.overrideController.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc00e22885c174317afdb9d1f35020ce
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 22100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
