+++ Assets/Prefabs/UI/Screens/LevelResultScreen/LevelResultScreenHide.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c5746adb728724df389f16bc1180ae0a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
