+++ Assets/Prefabs/UI/Screens/LevelResultScreen/LevelResultScreenShow.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6c50426d7c415442a9de4e3d0e69a85a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
