--- Assets/Scripts/Base/UI/HUD/UI_LevelResult.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 115e817dc24240c790aa548971f6a2c5
-timeCreated: 1727161108
