--- Assets/Scripts/Base/UI/HUD/UI_LevelUp.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: c58556b91bbc422a93b3e5ccd9bb70d3
-timeCreated: 1727794154
