--- Assets/Scripts/Base/UI/HUD/UI_NodeRewards.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 62aa9c5ead8d48688942aec7d45f7059
-timeCreated: 1727764947
